move validation methods to mapmodule util and use input in datecontrol
