+++ bundles/mapping/mapmodule/request/SetTimeRequestHandler.js
@@ -1,3 +1,4 @@
+import { validateDate, validateTime } from '../util/time';
 /**
  * @class Oskari.mapframework.bundle.mapmodule.request.SetTimeRequestHandler
  * Set time for map
@@ -20,12 +21,11 @@ Oskari.clazz.define(
             const date = request.getDate();
             const time = request.getTime();
             const year = request.getYear();
-
-            if (!this.validateTime(time)) {
+            if (!validateTime(time)) {
                 this.log.warn('Invalid time format. Valid format is HH:mm');
                 return;
             }
-            if (!this.validateDate(date, year)) {
+            if (!validateDate(date, year)) {
                 this.log.warn('Invalid date format. Valid format is D/M.');
                 return;
             }
@@ -33,30 +33,6 @@ Oskari.clazz.define(
             const timeArray = time.split(':');
             const dateObj = new Date(year, dateArray[1] - 1, dateArray[0], timeArray[0], timeArray[1]);
             this.mapModule.setTime(dateObj);
-        },
-
-        /**
-         * @method validateTime
-         * @return {Bool} true if valid time
-         */
-        validateTime: function (time) {
-            const regex = /^(0[0-9]|1[0-9]|2[0-3]|[0-9]):[0-5][0-9]$/;
-            return regex.test(time);
-        },
-
-        /**
-         * @method validateDate
-         * @return {Bool} true if valid date
-         */
-        validateDate: function (date, year) {
-            const matches = /^(0[1-9]|1[0-9]|2[0-9]|3[0-1]|[1-9])[/](0[1-9]|1[0-2]|[1-9])$/.exec(date);
-            if (matches === null) {
-                return false;
-            }
-            const d = parseInt(matches[1]);
-            const m = matches[2] - 1;
-            const dateObject = new Date(year, m, d);
-            return dateObject.getDate() === d && dateObject.getMonth() === m;
         }
     }, {
         /**
