+++ bundles/mapping/time-control-3d/view/TimeControl3d/Input.jsx
@@ -2,21 +2,22 @@ import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import { StyledIcon, StyledInput } from './styled';
 
-export const Input = ({ timeValue, changeHandler }) => {
-    const inputChangeTime = event => {
+export const Input = ({ value, changeHandler, iconType }) => {
+    const onChange = event => {
         const val = event.target.value;
         changeHandler(val);
     };
 
     return (
         <Fragment>
-            <StyledIcon type="clock-circle"/>
-            <StyledInput value={timeValue} onChange={inputChangeTime} />
+            <StyledIcon type={iconType}/>
+            <StyledInput value={value} onChange={onChange} />
         </Fragment>
     );
 };
 
 Input.propTypes = {
-    timeValue: PropTypes.string.isRequired,
-    changeHandler: PropTypes.func.isRequired
+    value: PropTypes.string.isRequired,
+    changeHandler: PropTypes.func.isRequired,
+    iconType: PropTypes.string.isRequired
 };
